==> AlgaeBloomForecast-20240926/00-StationCSV2shp.ipynb <==
# 将站点数据转换为shp文件

import chardet 
import pandas as pd
# 读取CSV文件
csv_file = '/root/Download/AlgaeBloomForecast/Stations2.csv'  # 替换为你的CSV文件路径

# 检测文件编码
with open(csv_file, 'rb') as f:
    result = chardet.detect(f.read())  # 读取文件内容，探测编码

print(result['encoding'])
# 使用探测到的编码读取文件
df = pd.read_csv(csv_file, encoding=result['encoding'])

# 将站点数据转换为shp文件

import geopandas as gpd
from shapely.geometry import Point

# 读取CSV文件
csv_file = '/root/Download/AlgaeBloomForecast/Stations2.csv'  # 替换为你的CSV文件路径

# 使用geopandas读取CSV数据，并指定编码格式为utf-8
gdf = gpd.read_file(csv_file, encoding=result['encoding'])

# 将经度和纬度转换为Point对象，创建几何列
gdf['geometry'] = gdf.apply(lambda x: Point(x['Longitude'], x['Latitude']), axis=1)

# 确保索引列是字符串类型
gdf['Cross Section'] = gdf['Cross Section'].astype(str)

# 将'断面'列设置为属性列
gdf = gdf.set_index('Cross Section')

# 指定坐标系统为WGS 84 (EPSG: 4326)
gdf.crs = 'EPSG:4326'

# 保存为Shapefile，并包含坐标投影信息
gdf.to_file('output_shapefile_BLOOMS2.shp')
