==> AlgaeBloomForecast-20240926/03-Station_merge_weather.ipynb <==
# 合并天气数据和站点数据
date,temp,oxygen,NH3,TP,TN,algae
2021/06/02,26.1875,6.6665,0.025,0.068275,1.07325,14400000.0
2021/06/03,25.881666666666664,6.6418333333333335,0.025116666666666666,0.06378333333333333,0.9151666666666666,10867091.666666666

...................
2024/05/19,23.868333333333336,8.204666666666666,0.04885,0.04328333333333333,1.294,21150000.0
2024/05/20,24.534,8.5298,0.05567999999999999,0.03828,1.4318,15640000.0
2024/05/21,24.595,8.073333333333332,0.05553333333333333,0.03763333333333333,1.6115000000000002,15316666.666666666



time,area,weather,max_temperature,min_temperature,aqi,aqiLevel,wind_direction,wind_power,aqiInfo
2021-01-01 00:00:00,无锡,晴,5,-4,56,2,西风,2级,良
2021-01-02 00:00:00,无锡,多云-晴,7,-2,71,2,东北风,2级,良

...................
2024-07-23 00:00:00,无锡,晴-多云,38,30,41,1,东南风,2级,优
2024-07-24 00:00:00,无锡,晴,36,28,23,1,东南风,2级,优



import pandas as pd

# 读取第一个CSV文件
df1 = pd.read_csv('/root/Download/AlgaeBloomForecast/smoothed_data.csv', parse_dates=['date'])

# 读取第二个CSV文件
df2 = pd.read_csv('/root/Download/AlgaeBloomForecast/sorted_weather_data.csv', parse_dates=['time'])

# 将 'time' 列重命名为 'date' 以便合并
df2 = df2.rename(columns={'time': 'date'})

# 确保两个DataFrame的日期列都只包含日期部分(不包含时间)
df1['date'] = df1['date'].dt.date
df2['date'] = df2['date'].dt.date

# 合并两个DataFrame，使用'date'列作为键
merged_df = pd.merge(df1, df2, on='date', how='outer')

# 按日期排序
merged_df = merged_df.sort_values('date')

# 将日期转换回字符串格式
merged_df['date'] = merged_df['date'].astype(str)

# 保存合并后的数据到新的CSV文件
merged_df.to_csv('merged_data.csv', index=False)

# 打印前几行结果
print(merged_df.head())
